Moved coordinate utils to `utils/coords.ts`
